TLB: resolve collisions
