--- rocket/src/util/package.scala
@@ -1,34 +1,0 @@
-// See LICENSE.SiFive for license details.
-// See LICENSE.Berkeley for license details.
-
-package org.chipsalliance.rocket
-
-import chisel3._
-import chisel3.util._
-import scala.math.min
-import scala.collection.{immutable, mutable}
-
-package object util {
-  def bitIndexes(x: BigInt, tail: Seq[Int] = Nil): Seq[Int] = {
-    require (x >= 0)
-    if (x == 0) {
-      tail.reverse
-    } else {
-      val lowest = x.lowestSetBit
-      bitIndexes(x.clearBit(lowest), lowest +: tail)
-    }
-  }
-
-  /** Similar to Seq.groupBy except this returns a Seq instead of a Map
-    * Useful for deterministic code generation
-    */
-  def groupByIntoSeq[A, K](xs: Seq[A])(f: A => K): immutable.Seq[(K, immutable.Seq[A])] = {
-    val map = mutable.LinkedHashMap.empty[K, mutable.ListBuffer[A]]
-    for (x <- xs) {
-      val key = f(x)
-      val l = map.getOrElseUpdate(key, mutable.ListBuffer.empty[A])
-      l += x
-    }
-    map.view.map({ case (k, vs) => k -> vs.toList }).toList
-  }
-}
